backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.0
pydantic==2.7.0
pydantic-settings==2.3.0
google-generativeai>=0.8.0
google-cloud-firestore>=2.16.0
google-cloud-storage>=2.17.0
google-cloud-texttospeech>=2.17.0
python-multipart==0.0.9
httpx==0.27.0
python-dotenv==1.0.1
